Don’t follow redirects
When proxying a request, one would not expect the proxy server to follow redirects and return the final destination, but rather to pass the redirect back to the client so _it_ can decide to follow it.

This change makes that happen, bringing micro-proxy more in line with how the now path alias feature works.

### Why?

There are many reasons why, but in our particular case we ran into this when our backend tried to set a cookie when redirecting (logging in), but this was swallowed by the proxy.